+++ src/arch/arm/fastmodel/CortexA76/FastModelCortexA76.py
@@ -31,9 +31,7 @@ from m5.objects.FastModel import (
 )
 from m5.objects.FastModelGIC import Gicv3CommsTargetSocket
 from m5.objects.Gic import ArmPPI
-from m5.objects.IntPin import IntSinkPin
 from m5.objects.Iris import IrisBaseCPU
-from m5.objects.ResetPort import ResetResponsePort
 from m5.objects.SystemC import SystemC_ScModule
 from m5.params import *
 from m5.proxy import *
