+++ src/arch/arm/fastmodel/GIC/FastModelGIC.py
@@ -40,8 +40,6 @@ from m5.objects.FastModel import (
     AmbaTargetSocket,
 )
 from m5.objects.Gic import BaseGic
-from m5.objects.IntPin import VectorIntSourcePin
-from m5.objects.ResetPort import ResetResponsePort
 from m5.objects.SystemC import SystemC_ScModule
 from m5.params import *
 from m5.SimObject import SimObject
