+++ src/arch/x86/X86LocalApic.py
@@ -39,7 +39,6 @@
 from m5.defines import buildEnv
 from m5.objects.BaseInterrupts import BaseInterrupts
 from m5.objects.ClockDomain import DerivedClockDomain
-from m5.objects.IntPin import IntSinkPin
 from m5.params import *
 from m5.proxy import *
