+++ src/dev/SConscript
@@ -35,9 +35,6 @@ Source('isa_fake.cc')
 Source('dma_device.cc')
 Source('dma_virt_device.cc')
 
-SimObject('IntPin.py', sim_objects=[])
-SimObject('ResetPort.py', sim_objects=[])
-
 DebugFlag('IntPin')
 DebugFlag('IsaFake')
 DebugFlag('DMA')
