+++ src/dev/riscv/RTC.py
@@ -33,7 +33,6 @@
 # (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
 # OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
-from m5.objects.IntPin import IntSourcePin
 from m5.params import *
 from m5.proxy import *
 from m5.SimObject import SimObject
