+++ src/dev/x86/Cmos.py
@@ -25,7 +25,6 @@
 # OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
 from m5.objects.Device import BasicPioDevice
-from m5.objects.IntPin import IntSourcePin
 from m5.params import *
 from m5.proxy import *
