+++ src/dev/x86/I8259.py
@@ -25,10 +25,6 @@
 # OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
 from m5.objects.Device import BasicPioDevice
-from m5.objects.IntPin import (
-    IntSourcePin,
-    VectorIntSinkPin,
-)
 from m5.params import *
 from m5.proxy import *
