+++ src/python/SConscript
@@ -380,6 +380,7 @@ PySource('m5.params', 'm5/params/deprecated_params.py')
 PySource('m5.params', 'm5/params/enum_params.py')
 PySource('m5.params', 'm5/params/network_params.py')
 PySource('m5.params', 'm5/params/null_params.py')
+PySource('m5.params', 'm5/params/pin_params.py')
 PySource('m5.params', 'm5/params/param_types.py')
 PySource('m5.params', 'm5/params/port_params.py')
 PySource('m5.params', 'm5/params/power_params.py')
